refactor: 프로필 패널 컴포넌트를 분리하고 라우팅 기반으로 패널을 관리하도록 수정
