+++ src/stores/sidebarStore.ts
@@ -4,19 +4,27 @@ interface SidebarState {
   isSecondaryOpen: boolean;
   selectedMenu: string | null;
   isLoggedIn: boolean;
-  profilePanelMode: 'login' | 'signup';
+  profilePanelMode:
+    | 'login'
+    | 'signup'
+    | 'profile'
+    | 'likes'
+    | 'followers'
+    | 'followings';
   openSecondarySidebar: (menu: string) => void;
   closeSecondarySidebar: () => void;
   toggleSecondarySidebar: (menu: string) => void;
   setLoginStatus: (status: boolean) => void;
-  setProfilePanelMode: (mode: 'login' | 'signup') => void;
+  setProfilePanelMode: (
+    mode: 'login' | 'signup' | 'profile' | 'likes' | 'followers' | 'followings'
+  ) => void;
 }
 
 export const useSidebarStore = create<SidebarState>((set, get) => ({
   isSecondaryOpen: false,
   selectedMenu: null,
   isLoggedIn: true, // 기본값: 로그인됨 (테스트용)
-  profilePanelMode: 'login', // 기본값: 로그인 모드
+  profilePanelMode: 'profile', // 기본값: 프로필 모드
   openSecondarySidebar: (menu: string) =>
     set({
       isSecondaryOpen: true,
@@ -48,7 +56,9 @@ export const useSidebarStore = create<SidebarState>((set, get) => ({
     set({
       isLoggedIn: status,
     }),
-  setProfilePanelMode: (mode: 'login' | 'signup') =>
+  setProfilePanelMode: (
+    mode: 'login' | 'signup' | 'profile' | 'likes' | 'followers' | 'followings'
+  ) =>
     set({
       profilePanelMode: mode,
     }),
